AP-14554: Validate custom workflow name "live".
This commit introduces functionality to the node dialogs of Capture
Workflow End and Workflow Writer that validates the name as the user
types. The settings cannot be saved if the input is invalid. No changes
to the configuration layout are made.

AP-14554 (Check if “Custom workflow name” in ”Capture Workflow End“ node has illegal characters )