AP-15066: Remove warning of invalid characters in custom workflow name
This commit removes the warning that was given if a custom workflow name
was provided by upstream nodes and that name contained characters that
had to be escaped. In combination with AP-14554, custom workflow names
will now always be validated upon input.

AP-15066 (Workflow Writer displays confusing warning on escaped default workflow name)